+++ Tests/Integration/Webhook/PaymentCaptureCompletedHandlerTest.php
@@ -9,15 +9,12 @@ declare(strict_types=1);
 
 namespace OxidSolutionCatalysts\PayPal\Tests\Integration\Webhook;
 
-use OxidEsales\Eshop\Application\Model\Order as EshopModelOrder;
 use OxidEsales\Eshop\Core\Registry as EshopRegistry;
 use OxidSolutionCatalysts\PayPal\Core\ServiceFactory;
 use OxidSolutionCatalysts\PayPal\Core\Webhook\Event as WebhookEvent;
 use OxidSolutionCatalysts\PayPal\Core\Webhook\Handler\PaymentCaptureCompletedHandler;
 use OxidSolutionCatalysts\PayPal\Exception\WebhookEventException;
-use OxidSolutionCatalysts\PayPal\Model\PayPalOrder as PayPalOrderModel;
 use OxidSolutionCatalysts\PayPal\Service\OrderRepository;
-use OxidSolutionCatalysts\PayPalApi\Model\Orders\Order as ApiOrderResponse;
 use OxidSolutionCatalysts\PayPalApi\Service\Orders as PayPalApiOrders;
 
 final class PaymentCaptureCompletedHandlerTest extends WebhookHandlerBaseTestCase
@@ -35,24 +32,6 @@ final class PaymentCaptureCompletedHandlerTest extends WebhookHandlerBaseTestCas
         $handler->handle($event);
     }
 
-    public function testEshopOrderNotFoundByPayPalOrderId(): void
-    {
-        $data = [
-            'resource' => [
-                'id' => self::TEST_RESOURCE_ID
-            ]
-        ];
-        $event = new WebhookEvent($data, static::WEBHOOK_EVENT);
-
-        $this->expectException(WebhookEventException::class);
-        $this->expectExceptionMessage(
-            WebhookEventException::byPayPalTransactionId(self::TEST_RESOURCE_ID)->getMessage()
-        );
-
-        $handler = oxNew(PaymentCaptureCompletedHandler::class);
-        $handler->handle($event);
-    }
-
     public function dataProviderWebhookEvent(): array
     {
         return [
@@ -61,26 +40,64 @@ final class PaymentCaptureCompletedHandlerTest extends WebhookHandlerBaseTestCas
             ],
             'api_v2' => [
                 'payment_capture_completed_v2.json'
-            ],
+            ]
         ];
     }
 
     /**
      * @dataProvider dataProviderWebhookEvent
      */
-    public function testPaymentCaptureCompleted(string $fixture): void
+    public function testPayPalTransactionIdWithoutPayPalOrderId(string $fixture): void
     {
         $data = $this->getRequestData($fixture);
+        $resourceId = $data['resource']['id'];
+        $event = new WebhookEvent($data, static::WEBHOOK_EVENT);
+
+        $loggerMock = $this->getPsrLoggerMock();
+        $loggerMock->expects($this->once())
+            ->method('debug')
+            ->with(
+                "Not enough information to handle PAYMENT.CAPTURE.COMPLETED with PayPal order_id '' and " .
+                "PayPal transaction id '" . $resourceId . "'"
+            );
+
+        EshopRegistry::set('logger', $loggerMock);
+
+        $handler = oxNew(PaymentCaptureCompletedHandler::class);
+        $handler->handle($event);
+    }
+
+    public function testEshopOrderNotFoundByPayPalOrderId(): void
+    {
+        $data = $this->getRequestData('payment_capture_completed_pui_v1.json');
+        $payPalOrderId = $data['resource']['supplementary_data']['related_ids']['order_id'];
+
+        $event = new WebhookEvent($data, static::WEBHOOK_EVENT);
+
+        $this->expectException(WebhookEventException::class);
+        $this->expectExceptionMessage(
+            WebhookEventException::byPayPalOrderId($payPalOrderId)->getMessage()
+        );
+
+        $handler = oxNew(PaymentCaptureCompletedHandler::class);
+        $handler->handle($event);
+    }
+
+    public function testPuiPaymentCaptureCompleted(): void
+    {
+        $data = $this->getRequestData('payment_capture_completed_pui_v1.json');
+        $payPalOrderId = $data['resource']['supplementary_data']['related_ids']['order_id'];
+
         $event = new WebhookEvent($data, static::WEBHOOK_EVENT);
 
         $orderMock = $this->prepareOrderMock();
-        $paypalOrderMock = $this->preparePayPalOrderMock($data['resource']['id']);
+        $paypalOrderMock = $this->preparePayPalOrderMock($orderMock->getId(), $payPalOrderId);
 
         $orderRepositoryMock = $this->getMockBuilder(OrderRepository::class)
             ->disableOriginalConstructor()
             ->getMock();
         $orderRepositoryMock->expects($this->once())
-            ->method('getShopOrderByPayPalTransactionId')
+            ->method('getShopOrderByPayPalOrderId')
             ->willReturn($orderMock);
         $orderRepositoryMock->expects($this->once())
             ->method('paypalOrderByOrderIdAndPayPalId')
@@ -89,9 +106,9 @@ final class PaymentCaptureCompletedHandlerTest extends WebhookHandlerBaseTestCas
         $orderServiceMock = $this->getMockBuilder(PayPalApiOrders::class)
             ->disableOriginalConstructor()
             ->getMock();
-        $orderServiceMock->expects($this->any())
+        $orderServiceMock->expects($this->once())
             ->method('showOrderDetails')
-            ->willReturn($this->getOrderDetails());
+            ->willReturn($this->getPuiOrderDetails());
 
         $serviceFactoryMock = $this->getMockBuilder(ServiceFactory::class)
             ->disableOriginalConstructor()
@@ -109,12 +126,5 @@ final class PaymentCaptureCompletedHandlerTest extends WebhookHandlerBaseTestCas
             ->method('getOrderRepository')
             ->willReturn($orderRepositoryMock);
         $handler->handle($event);
-    }
-
-    private function getOrderDetails(): ApiOrderResponse
-    {
-        $json = file_get_contents(__DIR__ . '/../../Fixtures/checkout_order_completed_with_pui.json');
-
-        return new ApiOrderResponse(json_decode($json, true));
     }
 }