+++ Tests/Integration/Webhook/WebhookHandlerBaseTestCase.php
@@ -12,6 +12,8 @@ namespace OxidSolutionCatalysts\PayPal\Tests\Integration\Webhook;
 use OxidEsales\Eshop\Application\Model\Order as EshopModelOrder;
 use OxidEsales\TestingLibrary\UnitTestCase;
 use OxidSolutionCatalysts\PayPal\Model\PayPalOrder as PayPalOrderModel;
+use OxidSolutionCatalysts\PayPalApi\Model\Orders\Order as ApiOrderResponse;
+use Psr\Log\LoggerInterface;
 
 class WebhookHandlerBaseTestCase extends UnitTestCase
 {
@@ -24,18 +26,26 @@ class WebhookHandlerBaseTestCase extends UnitTestCase
         return json_decode($json, true);
     }
 
-    protected function preparePaypalOrderMock(string $orderId): PayPalOrderModel
-    {
+    protected function preparePaypalOrderMock(
+        string $shopOrderId,
+        string $payPalOrderId = 'ppid',
+        string $transactionId = ''
+    ): PayPalOrderModel {
         $mock = $this->getMockBuilder(PayPalOrderModel::class)
             ->disableOriginalConstructor()
             ->getMock();
         $mock->expects($this->any())
             ->method('load')
-            ->with($orderId)
             ->willReturn(true);
         $mock->expects($this->any())
-            ->method('getId')
-            ->willReturn($orderId);
+            ->method('getPayPalOrderId')
+            ->willReturn($payPalOrderId);
+        $mock->expects($this->any())
+            ->method('getTransactionId')
+            ->willReturn($transactionId);
+        $mock->expects($this->any())
+            ->method('getShopOrderId')
+            ->willReturn($shopOrderId);
         $mock->expects($this->once())
             ->method('setStatus');
         $mock->expects($this->once())
@@ -46,7 +56,8 @@ class WebhookHandlerBaseTestCase extends UnitTestCase
 
     protected function prepareOrderMock(
         string $orderId = 'order_oxid',
-        string $methodName = 'markOrderPaid'
+        string $methodName = 'markOrderPaid',
+        string $expectCalls = 'once'
     ): EshopModelOrder {
         $mock = $this->getMockBuilder(EshopModelOrder::class)
             ->disableOriginalConstructor()
@@ -58,9 +69,38 @@ class WebhookHandlerBaseTestCase extends UnitTestCase
         $mock->expects($this->any())
             ->method('getId')
             ->willReturn($orderId);
-        $mock->expects($this->once())
+        $mock->expects($this->$expectCalls())
             ->method($methodName);
 
         return $mock;
+    }
+
+    protected function getPsrLoggerMock(): LoggerInterface
+    {
+        $psrLogger = $this->getMockBuilder(LoggerInterface::class)
+            ->disableOriginalConstructor()
+            ->setMethods(
+                [
+                    'emergency',
+                    'alert',
+                    'critical',
+                    'error',
+                    'warning',
+                    'notice',
+                    'info',
+                    'debug',
+                    'log'
+                ]
+            )
+            ->getMock();
+
+        return $psrLogger;
+    }
+
+    protected function getPuiOrderDetails(): ApiOrderResponse
+    {
+        $json = file_get_contents(__DIR__ . '/../../Fixtures/orderdetails_completed_with_pui.json');
+
+        return new ApiOrderResponse(json_decode($json, true));
     }
 }